requirements.txt
asgiref==3.3.1
beautifulsoup4==4.9.3
bootstrap4==0.1.0
Django==3.1.7
django-bootstrap4==2.3.1
django-cors-headers==3.7.0
djangorestframework==3.12.2
Pillow==8.1.0
psycopg2==2.8.6
pytz==2021.1
soupsieve==2.2
sqlparse==0.4.1
